optimize the build process - use ESM instead of CJS for the build script - let Sass output minified CSS files directly - use node-glob to dehardcode Sass input files - convert bare CSS files to equivalent Sass files, to unify the pipeline
