--- css/style-rtl.css
@@ -1,2 +1,0 @@
-@import "style-light-rtl.css";
-@import "style-dark-rtl.css" (prefers-color-scheme: dark);
