--- css/style.css
@@ -1,2 +1,0 @@
-@import "style-light.css";
-@import "style-dark.css" (prefers-color-scheme: dark);
