+++ sass/style-rtl.scss
@@ -1,0 +1,2 @@
+@import url("style-light-rtl.min.css");
+@import url("style-dark-rtl.min.css") (prefers-color-scheme: dark);
