+++ sass/style.scss
@@ -1,0 +1,2 @@
+@import url("style-light.min.css");
+@import url("style-dark.min.css") (prefers-color-scheme: dark);
